objectify download logic
